+++ app/src/main/java/net/justdave/nwsweatheralertswidget/widget/NWSWidgetPrefs.kt
@@ -17,28 +17,31 @@ private val Context.dataStore: DataStore<Preferences> by preferencesDataStore(na
 private fun areaKey(appWidgetId: Int) = stringPreferencesKey("appwidget_area_$appWidgetId")
 private fun zoneKey(appWidgetId: Int) = stringPreferencesKey("appwidget_zone_$appWidgetId")
 private fun titleKey(appWidgetId: Int) = stringPreferencesKey("appwidget_title_$appWidgetId")
+private fun themeKey(appWidgetId: Int) = stringPreferencesKey("appwidget_theme_$appWidgetId")
 private fun alertsKey(appWidgetId: Int) = stringPreferencesKey("appwidget_alerts_$appWidgetId")
 private fun updatedKey(appWidgetId: Int) = stringPreferencesKey("appwidget_updated_$appWidgetId")
 
-// Functions for saving and loading the widget's configuration (title, area, zone)
-suspend fun saveWidgetPrefs(context: Context, appWidgetId: Int, areaId: String, zoneId: String, title: String) {
+// Functions for saving and loading the widget's configuration (title, area, zone, theme)
+suspend fun saveWidgetPrefs(context: Context, appWidgetId: Int, areaId: String, zoneId: String, title: String, theme: String) {
     context.dataStore.edit {
         it[areaKey(appWidgetId)] = areaId
         it[zoneKey(appWidgetId)] = zoneId
         it[titleKey(appWidgetId)] = title
+        it[themeKey(appWidgetId)] = theme
     }
 }
 
 suspend fun loadWidgetPrefs(context: Context, appWidgetId: Int): Map<String, String?> {
-    return context.dataStore.data.map {
-        val alerts = it[alertsKey(appWidgetId)] ?: "[]"
-        mapOf(
-            "area" to it[areaKey(appWidgetId)],
-            "zone" to it[zoneKey(appWidgetId)],
-            "title" to it[titleKey(appWidgetId)],
-            "updated" to it[updatedKey(appWidgetId)],
-            "alert_count" to lenientJson.decodeFromString<List<NWSAlert>>(alerts).size.toString()
-        )
+    return context.dataStore.data.map { prefs ->
+        val map = mutableMapOf<String, String?>()
+        val alerts = prefs[alertsKey(appWidgetId)] ?: "[]"
+        map["area"] = prefs[areaKey(appWidgetId)]
+        map["zone"] = prefs[zoneKey(appWidgetId)]
+        map["title"] = prefs[titleKey(appWidgetId)]
+        map["theme"] = prefs[themeKey(appWidgetId)] ?: "semitransparent"
+        map["updated"] = prefs[updatedKey(appWidgetId)]
+        map["alert_count"] = lenientJson.decodeFromString<List<NWSAlert>>(alerts).size.toString()
+        map
     }.first()
 }
 
@@ -68,6 +71,7 @@ suspend fun deleteWidgetPrefs(context: Context, appWidgetId: Int) {
         it.remove(areaKey(appWidgetId))
         it.remove(zoneKey(appWidgetId))
         it.remove(titleKey(appWidgetId))
+        it.remove(themeKey(appWidgetId))
         it.remove(alertsKey(appWidgetId))
         it.remove(updatedKey(appWidgetId))
     }