+++ CHANGES.md
@@ -5,6 +5,7 @@
 * About dialog is readable in Dark mode now
 * Update interval is now 5 minutes instead of 15
 * Back button behavior on Android 12 from a widget-launched detail screen is now correct
+* Restores the semitransparent theme from the old version. But gives you the option per-widget to switch to the light and dark themes if you want them.
 
 ## Version 2.0 (Dec 13, 2025)